Backend/OCR/requirements.txt
pytesseract
pdf2image
PyMuPDF
Pillow
spacy
langchain
langchain-ollama
pydantic
python-dotenv
